Optimize editing the buffer
Try to only edit the minimum characters needed.  If we redraw the whole
line, and make it longer, all drawn regions on it shift to the right
(by the delta of the new length).

Either just replace the "rules" if the comment is already present, or
append a whole comment at the eol.